+++ changelog.md
@@ -3,6 +3,7 @@ Changelog
 
 ## 0.4.0
 
+* Added `getMetadata` method to fetch deploy metadata
 * Added `getVersions` and `removeVersion` methods. They will allow you to manage the deploys 
   currently on the device.
 * (fix) iOS deploys will now give a download error if the app goes into the background while